[VGNWB-366] Fix moving to another project
